refactor: 移动函数和常量
