+++ back_end/saolei/customranking/config.py
@@ -28,3 +28,4 @@ CUSTOM_PLUCK_MODES = {
     MS_TextChoices.Mode.NF,
     MS_TextChoices.Mode.RKC,
 }
+CUSTOM_PLUCK_CACHE_SIZE = 100
